libft avance petit a petit

+++ ft_strlcpy.c
@@ -3,13 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlcpy.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cjeannin <cjeannin@student.42.fr>          +#+  +:+       +#+        */
+/*   By: clement <clement@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/28 12:33:04 by cjeannin          #+#    #+#             */
-/*   Updated: 2025/11/03 10:42:47 by cjeannin         ###   ########.fr       */
+/*   Updated: 2025/11/03 22:59:36 by clement          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+// PAS OUF OUF SI SIZE = 0
 #include "libft.h"
 #include <bsd/string.h>
 #include <stdio.h>
@@ -17,7 +18,7 @@
 size_t	ft_strlcpy(char *dst, const char *src, size_t size)
 {
 	size_t	i;
-	int		j;
+	size_t	j;
 
 	j = 0;
 	while (src[j])
